feat(api): Add endpoint to allow user to resend confirmation email
A new API endpoint has been added to allow a user to resend a registration
confirmation to a user with the provided email.

Limitations:
For the MVP, this commit causes the API to rely on the Cognito service to rate
limit the requests rather than keep track of the requests itself.

closes #542