refactor(api): Prepare for EC2 deployment
Updated the minimum Python version to 3.10.
Migrated to pyproject.toml declarative project configuration which removes the
need for setup.py, requirements.txt, and test-requirements.txt.
tox.ini updated for minimum version 4.6.4.
Dockerfile updated to run `pip install` since requirements.txt has been deleted.
Dockerfile updated to remove the dependency on swagger-cli.

The reason for these changes is to prepare the code with the ability to build a
versioned wheel that can be deployed to a EC2 instance. The build will be done
by CI and the wheel uploaded to the EC2 instance.

The dependencies have been pinned to the latest version as of date.

A "variant" of this package named "dev" will install development dependencies
along with this application.

A "variant" of this package named "prod" will install "gunicorn" along with the
application so all of its dependencies can be installed in one install step.

Related #465