Create build-deploy-ec2.yml
This initial GitHub Workflow builds and uploads the Home Unite Us front-end application and back-end API web service to the EC2 instance currently used for the Home Unite Us project. 

In this commit, the workflow does not cause the uploaded packages to run. That will come after testing and refining this commit.

related #465