Fix path to openapi.yaml definition
The application's root path is used to form the path to the `openapi.yaml`
definition. This makes it so the path to the `openapi.yaml` file is relative to
the python module's location instead of the directory where the module is invoked.
For example, the module could be installed into `site-packages` and run from `/opt/HomeUniteUs`.

Related #539