API pyproject update setuptools version and pin dependencies
The version of the setuptools has been updated to the latest for this initial
migration to pyproject.

The prance and openapi-spec-validator versions have been pinned as a step
towards moving back to pinning dependency versions for reproducible
environments. A next step will be to reintroduce pinning transitive dependencies
and documenting a procedure for identifying core and transitive dependencies and
updating their versions.