fix: Ensure the API packages/modules are installed
The previous version caused the contents of the listed directories to be
searched for Python packages/modules. However, the alembic, configs, and
openapi_server packages are supposed to be seen as the packages to include in an
installation. This fix allows setuptools to discovery these packages while
excluding the tests package.