api: GitHub workflow fetch tags, API version extraction
The setuptools_scm tool derives the version from the latest tag and the distance
to that tag. This commit is a test to check if enabling the "fetch-tags" option
will allow setuptools_scm to derive the version.

This repository tags version starting with a 'v' but by default setuptools_scm
attempts to derive the version based on SemVer which does not start with a 'v'.
This commit modifies the default regex to extract the version from the tag
without the 'v'.