GitHub Workflow make build-deploy to EC2 workflow manually triggered
This commit modifies this GitHub workflow so that it is triggered manually. This
workflow will be listed under the GitHub Action's tab. Once it is selected, the
user will be able to select the "Run workflow" dropdown, select a branch or tag,
and then press the "Run workflow" button to run this workflow against the
selected branch or tag.

This change won't be visible in GitHub Actions until after it is merged into the
defaul branch (main).