fastapi-migration: Organization and tests
This commit contains files organized by functionality. It creates Python
packages via `__init__.py` in directories for packages that will be imported.

The `tests` directory was reorganized by test type end-to-end, integration, and
unit. All existing tests have been put into the integration directory.

The `housing service provider` has been renamed to `Housing Org` and the code
that updates this information is contained in the `tenant_housing_orgs`
directory in the `app` source code directory. The Repository for this has been
replaced with the `crud.py` file and the code for the controller has been
updated to implement HTTP method specifications more closely. The tests for this
controller have been updated to black box integration tests; meaning, it does
not inspect the database.