Add SQLAlchemy Event Store, Update Invite flow, Value Object
This commit includes an example of a "Value Object" named EmailAddress.
It represents a valid email address. Consumers of this object can expect that
the object is valid according to the data it contains such that the consumer
does not have to check if the email is valid.

An implementation of a SQLAlchemy-backed Event Store was also added along with a
unit test using a SQLite in-memory.

The Invite flow spec and code has been updated to include the idea of an invite
that is pending to be sent.